Extract the pre run checks into there own file
